Fixed the fighter battling

+++ MainEvent/Fighter.py
@@ -18,10 +18,10 @@ class Fighter:
     async def Level_Check(Self) -> bool:
         if Self.Data["Experience"] >= Self.Data["ExperienceRequired"]:
             Self.Data["Level"] += 1
-            Self.Data["Health"] += (Self.Data["Level"]*62)*1.5
-            Self.Data["Power"] += (Self.Data["Level"]*40)*1.5
-            Self.Data["Defense"] += (Self.Data["Level"]*40)*1.5
-            Self.Data["ExperienceRequired"] = (Self.Data["Level"]*225)*1.5
+            Self.Data["Health"] += int((Self.Data["Level"]*62)*1.5)
+            Self.Data["Power"] += int((Self.Data["Level"]*40)*1.5)
+            Self.Data["Defense"] += int((Self.Data["Level"]*40)*1.5)
+            Self.Data["ExperienceRequired"] = int((Self.Data["Level"]*225)*1.5)
             if Self.Data["Experience"] >= Self.Data["ExperienceRequired"]:
                 await Self.Level_Check()
             else:
